requirements.txt
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cryptography==1.9
defusedxml==0.5.0
Django==1.11.3
django-allauth==0.32.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
ecdsa==0.13
http-ece==1.0.4
idna==2.5
markdown2==2.3.4
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.0
py-vapid==1.2.5
pycparser==2.18
pyelliptic==1.5.7
python-decouple==3.0
python3-openid==3.1.0
pytz==2017.2
pywebpush==1.0.5
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.3
urllib3==1.21.1
